Frontend ready to connect with backend
