+++ front-end/react-crm/src/pages/auth/customAuth.tsx
@@ -1,49 +1,80 @@
-import React from 'react';
+import React, { useState } from 'react';
 import { Grid, Typography, Stack, TextField, Button } from '@mui/material';
+import { fetchData } from '../../components/FetchData';
+import { LoginUrl } from '../../services/ApiUrls';
+import { useNavigate } from 'react-router-dom';
 
-interface CustomAuthProps {
-    email: string;
-    setEmail: (value: string) => void;
-    password: string;
-    setPassword: (value: string) => void;
-    handleLogin: () => void;
-}
+const CustomAuth: React.FC = () => {
+    const navigate = useNavigate();
+    const [email, setEmail] = useState('');
+    const [password, setPassword] = useState('');
 
-const CustomAuth: React.FC<CustomAuthProps> = ({ email, setEmail, password, setPassword, handleLogin }) => {
+    const handleLogin = async () => {
+        try {
+            const res = await fetchData(
+                LoginUrl,
+                'POST',
+                JSON.stringify({ email, password }),
+                { 'Content-Type': 'application/json' }
+            );
+
+            console.log('Login Response:', res); // Log for debugging
+
+            if (res.access) {
+                localStorage.setItem('Token', `Bearer ${res.access}`);
+                navigate('/app');
+            } else {
+                console.error('Login failed:', res);
+                alert('Login failed. Please try again.');
+            }
+        } catch (error) {
+            console.error('Error during login:', error);
+            alert('An unexpected error occurred during login. Please try again later.');
+        }
+    };
+
     return (
-        <Grid container item sx={{ mt: 4, width: '100%' }} direction="column" alignItems="center">
-            <Typography variant="body2" sx={{ mb: 1 }}>
-                Or you can login with your email and password.
-            </Typography>
-            <Stack spacing={2} sx={{ width: '100%' }}>
-                <TextField
-                    label="Email"
-                    variant="outlined"
-                    size="small"
-                    fullWidth
-                    value={email}
-                    onChange={(e) => setEmail(e.target.value)}
-                />
-                <TextField
-                    label="Password"
-                    type="password"
-                    variant="outlined"
-                    size="small"
-                    fullWidth
-                    value={password}
-                    onChange={(e) => setPassword(e.target.value)}
-                />
-                <Button
-                    variant="contained"
-                    color="primary"
-                    fullWidth
-                    onClick={handleLogin}
-                    sx={{ mt: 2, backgroundColor: '#3e4b68', '&:hover': { backgroundColor: '#2d3748' }}}
-                >
-                    Login
-                </Button>
-            </Stack>
-        </Grid>
+        <form
+            onSubmit={(e) => {
+                e.preventDefault();
+                handleLogin();
+            }}
+            style={{ width: '100%' }}
+        >
+            <Grid container item sx={{ mt: 4, width: '100%' }} direction="column" alignItems="center">
+                <Typography variant="body2" sx={{ mb: 1 }}>
+                    You can login with your email and password.
+                </Typography>
+                <Stack spacing={2} sx={{ width: '100%' }}>
+                    <TextField
+                        label="Email"
+                        variant="outlined"
+                        size="small"
+                        fullWidth
+                        value={email}
+                        onChange={(e) => setEmail(e.target.value)}
+                    />
+                    <TextField
+                        label="Password"
+                        type="password"
+                        variant="outlined"
+                        size="small"
+                        fullWidth
+                        value={password}
+                        onChange={(e) => setPassword(e.target.value)}
+                    />
+                    <Button
+                        variant="contained"
+                        color="primary"
+                        type="submit"
+                        fullWidth
+                        sx={{ mt: 2, backgroundColor: '#1976d2', '&:hover': { backgroundColor: '#0f569c' }}}
+                    >
+                        Login
+                    </Button>
+                </Stack>
+            </Grid>
+        </form>
     );
 };
