+++ front-end/react-crm/src/services/ApiUrls.tsx
@@ -1,6 +1,6 @@
 // authurl
 export const SERVER = process.env.REACT_APP_API_BASE_URL
-export const LoginUrl = 'api/token'
+export const LoginUrl = '/api/api/token'
 // export const LoginUrl = 'auth/login'
 // export const RegisterUrl = 'auth/register'
 export const ForgotPasswordUrl = 'auth/forgot-password'
